requirements.txt
aws-cdk-lib==2.27.0
black==22.3.0
constructs==10.1.31
fastapi==0.75.1
fire==0.4.0
pylint==2.13.8
pytest==7.1.2
pytest-cov==3.0.0
textblob==0.17.1
uvicorn==0.17.6
wikipedia==1.4.0